requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
requests==2.32.3
python-jose[cryptography]==3.3.0
google-cloud-firestore==2.19.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
